Prevent uploading a file that we've already uploaded
Store the filename and photoId of each file that we upload into
.rodeo-uploaded-files.json and on upload, if the file to be uploaded is
in this list, then do not upload it again.

By storing the photoId along with the filename, we can present a link to
the user for them to view the image on Flickr if it was previously
uploaded.

Add Contains() to internal

Contains() determins if a string exists in an array of strings.

Prevent uploading a file that we've already uploaded

Store the filename of each file that we upload into .rodeo-uploaded-files.json
and on upload, if the file to be uploaded is in this list, then do not
upload it again.

Remove Contains() from internal.strings

No longer needed.

Store photoId along with filename in uploaded files record

By storing the photoId along with the filename, we can present a link to
the user for them to view the image on Flickr.