Refactor GetFlickrClient() and GetAlbums() to flickr.go
As we need to list albums in more than one command, we refactor the
ability to list albums from the `listalbums` command to
`internal/flickr.go` as `GetAlbums()`. To do this, we also need to
extract `GetFlickrClient()` from the `upload` command into
`internal/flickr.go` too.

This simplifies that `listalbums` command.